importer/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
geojson==2.5.0
idna==3.4
importlib-metadata==5.0.0
Pillow==9.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
tabulate==0.9.0
typing_extensions==4.4.0
urllib3==1.26.12
zipp==3.9.0
